Move some stuff to CommandStringHelper
